Clean up Detection and Segmentation code
